+++ src/app/components/acquisition/CloudSelector.jsx
