+++ src/app/components/acquisition/SatelliteCard.jsx
