+++ src/app/components/acquisition/Slider.jsx
