+++ src/app/components/acquisition/Tick.jsx
