+++ src/app/components/core/Footer.jsx
