+++ src/app/components/core/index.js
@@ -1,0 +1,6 @@
+export { default as ActivityIndicator } from './ActivityIndicator'
+export { default as DropdownButton } from './DropdownButton'
+export { default as Footer } from './Footer'
+export { default as NavigationBar } from './NavigationBar'
+export { default as Scroller } from './Scroller'
+export { default as User } from './User'
