+++ src/app/components/index.js
@@ -1,3 +1,4 @@
-export { default as Scroller } from './Scroller'
 export * from './snacks'
 export * from './visualization'
+export * from './map'
+export * from './core'
