+++ src/app/components/map/ImageChooserForm.jsx
