+++ src/app/components/map/LoadedImagesAccordion.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { useSelector } from 'react-redux'
 import { MuiThemeProvider } from '@material-ui/core/styles'
 
-import ImageViewer from '../components/ImageViewer'
-import { lessPaddedExpansion } from '../theme'
+import ImageViewer from './ImageViewer'
+import { lessPaddedExpansion } from '../../../theme'
 
 const LoadedImagesAccordion = ({ }) => {
   const images = useSelector(state => state.imagery.images)
