+++ src/app/components/map/ShapeList/index.js
@@ -1,0 +1,1 @@
+export { default as ShapeList } from './ShapeList'
