+++ src/app/components/map/index.js
@@ -1,0 +1,11 @@
+export * from './ShapeList'
+export { default as ActionList } from './ActionList'
+export { default as FeatureInfo } from './FeatureInfo'
+export { default as GoogleMap } from './GoogleMap'
+export { default as ImageChooserCard } from './ImageChooserCard'
+export { default as ImageChooserForm } from './ImageChooserForm'
+export { default as ImageViewer } from './ImageViewer'
+export { default as LayerActions } from './LayerActions'
+export { default as LayerViewer } from './LayerViewer'
+export { default as LoadedImagesAccordion } from './LoadedImagesAccordion'
+export { default as OpacityControl } from './OpacityControl'
