+++ src/app/components/visualization/SatelliteImageThumbnail.jsx
