--- src/app/components/visualization/ShapeList/index.js
@@ -1,1 +1,0 @@
-export { default } from './ShapeList'
