+++ src/app/components/visualization/index.js
@@ -1,1 +1,5 @@
-export { default as ShapeList } from './ShapeList'
+export { default as EvolutionChart } from './EvolutionChart'
+export { default as ImageTable } from './ImageTable'
+export { default as SatelliteImageThumbnail } from './SatelliteImageThumbnail'
+export { default as TransectDataTable } from './TransectDataTable'
+export { default as TransectEvolution } from './TransectEvolution'
