+++ src/app/pages/AcquisitionPage/index.jsx
@@ -7,11 +7,11 @@ import { useTranslation } from 'react-i18next'
 import { makeStyles } from '@material-ui/core/styles'
 import { Divider, Grid, Paper, Step, StepLabel, Stepper, Typography } from '@material-ui/core'
 
-import SatelliteChooser from '../../../containers/SatelliteChooser'
-import AOIChooser from '../../../containers/AOIChooser'
-import PeriodChooser from '../../../containers/PeriodChooser'
-import ImageListRefiner from '../../../containers/ImageListRefiner'
-import Footer from '../../../components/Footer'
+import SatelliteChooser from '../../components/acquisition/SatelliteChooser'
+import AOIChooser from '../../components/acquisition/AOIChooser'
+import PeriodChooser from '../../components/acquisition/PeriodChooser'
+import ImageListRefiner from '../../components/acquisition/ImageListRefiner'
+import Footer from '../../components/core/Footer'
 
 import { getAcquisitionParameters } from '../../../selectors'
 import { Actions as Auth } from '../../../store/ducks/auth'
